requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
kornia
natsort
psutil
opencv-python
SimpleITK